Add files via upload
Hi,
Its difficult to find a Barchart or simple  graphic expemple for MacOS written in Swift Cocoa.  I am not an expert but I have decide to share what I have found.
This program is a simple Barchart exemple build  with rectangle and text.
If you have a choice, using SwiftUI would be simpler.